+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-- Recommended minimum Gradle version is now 7.4.1.
+- Recommended minimum Gradle version is now 7.5.1.
 
 ## [0.8.0] - 2022-01-17
 ### Changed