+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.10.0] - 2023-12-01
 ### Changed
 - Recommended minimum Gradle version is now 8.5.
 - Update ZAP API client to version 1.13.0.
@@ -226,7 +226,7 @@ Provided by the plugin:
  with a `ZapVersions.xml` from an add-on.
 
 
-[Unreleased]: https://github.com/zaproxy/gradle-plugin-add-on/compare/v0.9.0...HEAD
+[0.10.0]: https://github.com/zaproxy/gradle-plugin-add-on/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/zaproxy/gradle-plugin-add-on/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/zaproxy/gradle-plugin-add-on/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/zaproxy/gradle-plugin-add-on/compare/v0.6.0...v0.7.0